requirements.txt
streamlit
Pillow
rembg
mediapipe
numpy